checks: Skip LFS checks when deleting refs
The LFS check asserts that for a given reference update, all newly
reachable LFS pointers refer to existing LFS objects. It is trivial to
see that when a reference gets deleted, no new git objects can become
reachable and thus its not possible for the deletion to introduce new
LFS pointers. There already is a spec which asserts that integrity
checks are not executed in case the reference is being deleted. But the
spec itself is wrong: it uses an empty newrev, while git uses the
all-zeroes object name to identify deletions.

Fix both the test and the uncovered bug that LFS integrity checks
actually do run for deletions.