Add pinning test for current implementation
Here we set up a [pinning test](http://wiki.c2.com/?PinningTests) to
lock down the complete behaviour of the current implementation of the
Security Configuration page.

Because the state of the configuration table is completely determined by
the `data-` attributes of the mount element, it's easy to mock various
scenarios, just by copying the returned markup from the page of
a running instance, e.g., production.

Three projects from production are used to drive the tests:

1. `gitlab-org/gitlab-ui` (this has some manually enabled scanners)
1. `markrian/qr` (this has _no_ enabled scanners)
1. `markrian-test/test-public-project` (this has some scanners enabled
   via Auto DevOps)

Since these test establish the current behaviour, the tests pass by
definition.