Switch to new implementation
This commit switches to the new implementation of the table. It's best
to view this commit by ignoring whitespace changes, since the indenting
level of some elements has changed, but the overall structure and
strings have not. On the command line, you can do something like:

    git show <this commit> --ignore-all-space

Or by viewing this commit in GitLab and appending `?w=1` to the URL.