Geo: Fix syncing remote stored Blobs
When direct download is used.

`http.rb`'s `follow` method forwards headers from the original request
with the redirect, and it's normal for the remote store to respond 400
Bad Request if it receives an unexpected header like:
`Authorization: Gl-Geo ...`. So here we follow redirects manually.

Also, copy this timeout fix at the same time since it's in the same
lines of code:
https://gitlab.com/gitlab-org/gitlab/-/commit/15802cc45dfa88bda5b6b388a7e8fd16f450c917?merge_request_iid=36012