Updates to CI_CONFIG_PATH and CI_PROJECT_CONFIG_PATH variables
CI_PROJECT_CONFIG_PATH is same as CI_CONFIG_PATH.
This MR moves CI_CONFIG_PATH as a Project variable
(so it can use it in the include section).
This is necessary since the include section
can resolve on pre-defined project variables.
Removes support for CI_PROJECT_CONFIG_PATH to avoid duplication
of variables that have the same behavior