Ensure issue closing pattern example is valid
The suggested addition to gitlab.rb produces a different regular
expression than expected because it is encoded as a Ruby string, so
escapes produce different results. This change writes the regular
expression in Ruby's Regexp syntax and then converts it to the source
string to preserve the syntax of the regular expression with the
following YAML example.