Use actual file IDs to retrieve reviews
It's not very robust to retrieve a value by its position in an array.

Instead, because the files all have unique IDs we return an
object where each key is a file ID and it's value is whether that
file has been marked as reviewed.

This fixes the issue where the list of files in single-file mode
always has a length of 1, which means the fetched review is
always the first file in the MR.