Add link to merge request in email
The link to merge requests in the emails
about new merge requests being created
are missing in the entire content until the
footer. This causes a lot of scroll just to
visit the merge request page on the UI.

The text version of the same email includes
a link on the opening line, so the HTML
version should do the same.