Default note *_access_level to DISABLED instead of nil
This relates to some more possible edge cases we may have in our data.
Since we are running a migration to backfill all these settings that
will run forever unless a value is set it is better to always have some
fallback that isn't `nil` as the migration will never finish.

In this case we can be sure that anything that is required to have a
project but doesn't for whatever reason shouldn't show up in searches
and therefore it's safe to just set `*_access_level = DISABLED`.