BulkImports: Add fields to track concurrent work
Add a way to track the status of each BulkImport's pipeline
individually. The end goal is to enable BulkImports to run in smaller
and more concurrent jobs. The pipeline status/progress will be tracked
on the `bulk_import_trackers` table, using `BulkImports::Tracker` model.

More information on:
- https://gitlab.com/gitlab-org/gitlab/-/issues/323382

The next steps
- https://gitlab.com/gitlab-org/gitlab/-/issues/323384
  will be to run each pipeline on its own background job.