Do not mark bot users as pending approval when User Cap is enabled
When User cap is enabled we mark all newly created users as pending
approval, even those that are not humans. This prevents creation of
project access tokens - at the time when we try to create the token the
associated bot users is not approved.

This updates `Users::BuildService#set_pending_approval_state` to mark
users as pending approval only if they are `human?`.

See https://gitlab.com/gitlab-org/gitlab/-/issues/323191.