Add /internal/geo_proxy endpoint
Which is intended for Workhorse to use to determine whether it is in a
Geo secondary site. If so, then it should proxy requests to the Geo
primary site via the URL returned by this endpoint.

When the feature flag "geo_secondary_proxy" is disabled, this endpoint
returns empty data even if this is a Geo secondary site.

Part of https://gitlab.com/groups/gitlab-org/-/epics/1528