Wiki: #find_file defaults to HEAD over master
Gitaly will find Wiki files is requested to do so, and by default if no
revision is set, it will set the revision to `master`. Given the Git
community is moving towards `main` instead, this change sets the default
to HEAD.

This works as GitLab stores repositories in their bare format, and
effectively there's no HEAD. Git itself really likes there to be a HEAD,
and this thus points to the default branch name. Nowadays usually
`refs/heads/master`, in the future probably: `refs/heads/main`.