Fix crashing in-product-marketing email worker
This is a temporary fix for the in-product-marketing-email worker.
It is a bug that consists of multiple layers. We expect to only have
root namespaces in `onboarding_progresses`. However, groups can be moved
to sub-groups. Threfore we also have sub-groups that would receieve the
emails.

This alone would not be an issue. But we need to call `.root_ancestor`
for these sub-groups and it doesn't play well together with
`each_batches` when there is an other JOIN condition on the namespace.
It currently generates a non-working query that fails.

To prevent this for now we do not send emails to sub-groups as a
temporary solution. In the long run, we need to remove all sub-groups
fron `onboarding_progresses` and fix the issue of calling
`.root_ancestor` in an `each_batches` call.