Fix application settings helpers not being loaded in FOSS
In https://gitlab.com/gitlab-org/gitlab/-/issues/324279, we saw a
failure only in the FOSS build because EE includes the
`ApplicationSettingsHelper` in the EE routing helper but FOSS does not.

Since the loading of this helper isn't really EE-specific, we ensure all
the includes are done in both EE and FOSS. This should eliminate subtle
differences between CE and EE code.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/324279