Replace generate command with direct Thor calls
This replaces `generate` calls with direct calls to Thor API to run
metric definition generators. This provides visible coupling to the
generator in question and resolves problems with stubbing constants in
RSpec, which were caused with `generate` calling the
UsageMetricDefinitionGenerator through a system call, spawning a new
process and circumventing RSpec's stubbing.