Create Apollo Link for Captcha handling
This implementation of an Apollo Link is similar to our Axios
Captcha interceptor. The Link checks whether a top-level error in the
GraphQL response hints towards the need of solving a captcha. If the
captcha is solved correctly, the mutation will be send again, but this
time the solution to the CAPTCHA is added to the HTTP headers.