Link squashed commits using the changelog API
When generating changelogs using the API, we now link to merge requests
for squashed commits. Prior to this commit, when a merge request was
squashed into a target branch, we weren't able to link back to that
merge request in a generated changelog.

See https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1551 for
more information.