Docs: rework coverage tools regex examples
An attempt to fix Markdown escaping in coverage regex example for nyc
ended up in a tricky scenario where either the rendered GitLab help page
or the html on docs.gitlab.com showed the regex incorrectly.

To work this around, we replace the Markdown table with a list of the tools
and the sample regular expressions.