Optimize IssuesFinder redundent confidential check when assigned
This query is sub-optimal and is frequently reported in the top 15 in
total time taken due to the fact that this is used to count the number
of assigned open issues on header bar which is loaded on every page
load.

Since a common use case is filtering by assigned to self (ie. the count
of assigned issues) we can optimize this query by not bothering to check
for confidentiality as it doesn't affect the outcome as you can see
assigned issues regardless of confidentiality.