This adds a context around expiring members
The context will be different depending on the type of member that is
being expired:

For a `GroupMember` the `source` relation will be a group, so the
`meta.root_namespace` will the root ancestor of that group.

For a `ProjectMember` the `source` relation will be a project. So the
`meta.project` will be set to that project, and the
`meta.root_namespace` will become the root namespace of that project.

The `meta.user` will always be set to the username of the user for
which the membership is being expired.