Hide pipeline status in pipeline editor for new projects
When a user opens a new project and uses the pipeline editor to create a
new CI configuration file, the pipeline status should be hidden since
there are no triggered pipelines yet for the project. It will be visible
once the user creates their first commit (and thus, triggers their first
pipeline).