Add feature flag to apply upload API size enforcement
This adds the `enforce_max_attachment_size_upload_api` feature flag to
enable enforcement of max attachment size in the upload API.  If the
feature is not enabled, we limit to 1 GB by
default. `GITLAB_UPLOAD_API_ALLOWLIST` can be set in the environment to
allow certain project IDs to bypass enforcement.