Split up the int column conversion helpers
Split the `initialize_conversion_of_integer_to_bigint` migration helper
into two pieces. The first remains as it is, which creates the temporary
column and installs the trigger.

The new second method, `backfill_conversion_of_integer_to_bigint`,
handles the next step of the migration, which is creating the batched
background migration that will backfill the existing data in the
temporary column.