Invalidate todo counter cache to avoid N+1 SQL queries
When deleting an issuable (Issue or MergeRequest), todo counters
(done and pending) needs to show updated counts. We are caching
them in Redis for a certain duration (as of this writing, it's 20
mins).

Before this change, we force the cache update which results to
N+1 SQL issues because it's going to execute a COUNT query for
each user of associated todos of the deleted issuable.

The fix for that is to invalidate the cache of the said counters
so we will only execute the COUNT query when a user views the
counter (which will be in separate requests).