Add message for repository backup skip
The backup procedure for repositories will print `[SKIPPED]` for any
repository it finds empty, or non-existent.

To an uninformed user reading just the terminal, the reason for the skip
isn't clear, and it may also be alarming to them.

This commonly applies to the .wiki or .design repos.

This change alters the message to suggest the repository was empty.