Fix test history badge alignment
Put the badge back inside the button. If it's not
there, then the gl-badge stretches to fill the
space if the issue name takes up multiple lines.
I could instead use `gl-display-items-flex-start`
but that caused the issue name to be aligned to
the top of the badge instead of the text in the
badge.