BulkImports: Remove epics author finding duplication
Mapping users/author of imported relations is a recurrent work to be
done. Therefore, the
BulkImports::Common::Transformers::UserReferenceTransformer was created
to centralized how it's done.

This commit adds the
BulkImports::Common::Transformers::UserReferenceTransformer as a
transformer in
EE::BulkImports::Groups::Pipelines::EpicsPipeline and removes the
duplicated logic from the
EE::BulkImports::Groups::Transformers::EpicAttributesTransformer