Remove add issues modal code from issue boards
This feature has been disabled since 13.6 with a feature flag. Seems
like we have not gotten any feedback that we should enable it, so let's
remove it from the codebase.

One motivation for removing it: It uses the Vue runtime template
compiler. Removing this will bring us one step closer to shipping Vue
without the template compiler which will bring a performance improvement
across all pages.

https://gitlab.com/gitlab-org/gitlab/-/issues/205169
https://gitlab.com/gitlab-org/gitlab/-/issues/292803