Make sure we're looking at the right package_file
If we create a new package, the factory will also create a corresponding
package file, which will share the same name as the package file we're
creating explicitly, and it only has file_sha1 defined therefore if
PostgreSQL happens to load the package file sharing the same name from
the package, then it will fail.

This make sure we don't create a new package thus avoid having an extra
package file which will conflict with the package file we're testing
here.