Support automatic transitions of Jira issues
Add a new setting to the Jira integration to attempt an "automatic"
transition of Jira issues, by checking the available workflow
transitions on the issue and using the first one with a status
category of "Done".

Automatic transitions are enabled by setting the new DB column
`jira_issue_transition_automatic` to true, which takes precedence over
the existing `jira_issue_transition_id` column.

To make this behaviour opt-in, we also introduce an explicit checkbox
to enable issue transitions. The state of this checkbox is handled
entirely on the frontend, based on the values in
`jira_issue_transition_id` and `jira_issue_transition_automatic`.