Ensure a project iid is set before transitioning on pipeline error
Issue: gitlab.com/gitlab-org/gitlab/-/issues/325861
MR: gitlab.com/gitlab-org/gitlab/-/merge_requests/57783

We need ensure_project_iid to run without any parent transaction,
otherwise the InternalId table for that scope, in our case project,
will be locked for the entire transaction. If we error before the seed
chain method is called, then the pipeline will not have an iid set,
and the transition to drop! will have a transaction, so when we
ensure the project iid, it will be merged into the transition
transaction.