UpdateSubmodule: Update error expectation in test
This change is made to unblock both the release of a new Gitaly as well
as the rollout of the UpdateSubmodule RPC. This RPC was ported to Go,
and with that an error message now is more precise. Sadly this breaks
the GitLab test suite.

Assertions should not be made against user facing error messages I
think, though we keep that test assertion and just update it now.