Add runner type label to queue retrieval duration histogram
This commit adds a runner_type label to the Prometheus histogram we use
to record the time it takes to retrieve a builds queue from the
database.

We also add this label to the histogram we use to track duration of the
entire process of processing a queue for a given runner.