Replace pages migraiton feature flag with the ENV variable
Currently we mark pages projects as not deployed if
feature flag is enabled.

We want to give users control over this behavior and
I'm afraid of just enabling feature flag for everyone.

In worst case scenario enabling this feature flag and running
migration may result in all pages projects marked as not deployed.
(e.g. if users run task on the node without NFS share attached)

So the idea is:
* we'll run the migration in background without this behavior enabled
* users will manually run the task and see all the errors logged
* if the decide that these errors are legit, they can mark project
as "pages not deployed" by re-running rake task with the ENV variable
enabled