Move commit neighbor buttons to sticky MR controls
When reviewing an MR commit by commit, there are `Prev` and `Next`
buttons next to the commit message to jump to the neighbor
commits. These are helpful, but they slide outside of view when
scrolling to the bottom.

These buttons were initially introduced in [1].

This changes moves these buttons to the sticky Merge Request controls
bar. This bar remains visible at all times, making the neighbor commit
buttons accessible even when scrolled to the bottom.

1. ee92100397f (Add buttons to the Commit Item to navigate among MR
   commits, 2020-05-12)