Refactoring Sidebar Project Panel
In this commit we start the refactor of the project sidebar.

Instead of the current approach with all the logic in views and helpers,
we're moving to a model based structure where each menu model contains
its logic, visibility access levels, and all the necessary info.

Nevertheless, the migration is tricky because we cannot just replace
the whole sidebar (that would be a 130 files MR). In order to
make it incremental we still render the old views. Then, when we
add a new menu to the new structure we'll remove that menu
from the menu, until we migrate all menus and we can remove the view
entirely.