Handle assignee changes side effects asynchronously
This is to reduce the number of queries that get executed by
running `MergeRequests::UpdateService`.

This is based on the solution made in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/57523.

Added a new worker/service to extend the capability of the worker
added in that merge request. It now accepts a `execute_hooks`
option so it can be re-used by both `UpdateService` and
`UpdateAssigneesService`.