Render experiment variant in active panel
This is a fix for a regression that was introduced during the final
cycle of a review for a previous MR (linked below). The regression was
that the experiment variant was no longer occuring in a subcomponent.

This MR backfills a test to ensure that the experiment is experienced
in the subcomponent by decorating data before it's passed down.

previous MR:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/55818

part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/285286