IPython KaTeX rendering of comparison operators for markdown
When rendering IPython notebook markdown cells the formulas are passed
to Katex. Before that the raw string is passed to marked.js to handle
HTML forbidden characters. Two of these characters are the
less-than-operator `<` and the greater-than-operator `>`. These
characters are often used in math formulas and the replacements `&lt;`
and `&gt;` are not well received by Katex.

For the tests use async/await syntax, and check length with
`toHaveLength()` to be more explicit.

Fixes: https://gitlab.com/gitlab-org/gitlab/-/issues/327450