Highlight non-recursion of `artifacts:exclude`
The file-matching syntax of `artifacts:paths` and `artifacts:exclude`
seem (and are) similar, but `exclude` differs in that it does not
include directory contents recursively. This can lead to confusion when
attempting to exclude the contents of a subdirectory by matching that
subdirectory's path. Explain this caveat in the docs and add a relevant
example.

See https://gitlab.com/gitlab-org/gitlab/-/issues/15122#note_524690377
for an example of this issue in the wild.