Use actual activated at value for license type
Prior to this change, the value for activated at in the GraphQL
Types::Admin::CloudLicenses::LicenseType class was using created at
since the field didn't exist before. Now that the field exists in the
Gitlab::License we can use the actual value and implement a fallback to
created at for older licenses.