Log all API uploads that exceed max attachment size
Previously if the project were on the `GITLAB_UPLOAD_API_ALLOWLIST`
exemption list we would not log a message. However, it is still useful
to track these exceptions so that we know that they are happening. We
add a `upload_allowed` boolean field to make it possible to filter these
exempted uploads.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/325788