Bump minimum required Go version for Workhorse to 1.15
We're currently in the process of moving to Go 1.16 across all
distributions, with a minimum required Go version of 1.15. Workhorse is
still using Go 1.13 as the minimum required Go version, which doesn't
make a lot of sense when other components already require a newer
version anyway.

Bump the minimum required version of Go to 1.15 and move over the
default image in CI to 1.16.