Upgrade staticcheck tool to fix false positive
With the upgrade to Go 1.16, the staticcheck analyzer complains about a
potential nil pointer access in "cmd/gitlab-zip-cat/main.go". This is a
false positive: if the pointer was a `nil` pointer, we'd already have
called `os.Exit()`.

This issue has already been fixed upstream by improving analysis on
functions which do not ever return. So let's upgrade the staticcheck
executable to a newer release.