Reduce timeouts on SearchController#count to 5s
The current timeout defaults to 15s. Per
https://gitlab.com/gitlab-org/gitlab/-/issues/327106 this endpoint
consumes a considerable amount of DB time. Since this endpoint is not as
essential as the primary task the user is performing (since it is used
to generate tab counts for other tabs you aren't searching) it makes
sense to reduce the timeout so slower counts don't consume too much
resources.