Work around nested attributes
In Ci::Processable, we define `accepts_nested_attributes_for :needs`.
Unfortunately, this does not exclude columns ignored on `Ci::BuildNeed`.

There is a database trigger in place, that always keeps the
`build_id_convert_to_bigint` column in sync with `build_id`. Hence, a
copy of a build (e.g. when retrying) will have a different value for
this column compared to before.

This should be a temporary workaround while we maintain that column
through database trigger.