Fix creating an MR when a project approval rule has no approvers
Without this change, we assume the regular approval rule is actually an
"any_approver" approval rule. An MR can only have one of those, so we
get a constraint violation error when the regular approval rule is
converted.

It's odd that these rules are entirely valid in the backend, but making
them invalid at this stage feels a bit risky.

Changelog: fixed