Remove HipChat service and records
Hipchat was discontinued by Atlassian in 2018, and support reached
end-of-life in June of 2020.

We originally planned to remove it in 14.0, but due to the issue with
mimemagic dependency
https://gitlab.com/gitlab-org/gitlab/-/issues/325851 we had to make it
non-operational in 13.11. The gem and some code have been removed in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/57434 and
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/57556.

This continue removing more code and deletes all the HipChat service
records from the database.