Change JiraTrackerData#deployment_type based on URL
We have around 90k records for this table where deployment_type is
unknown. Our logic for JIRA integration started to diverge depending on
whether the user is trying to integrate with JIRA Cloud or JIRA Server.

This migration will try determining what is the deployment type for
every entry in jira_tracker_data table based on the URL.

The logic is simple – if URL ends in `atlassian.net` then most likely
we're dealing with JIRA Cloud, otherwise it's JIRA Server.

Changelog: other