Introduce new service to handle add_spent_time for MRs
This new service creates a timelog from the given duration, creates a
system note, tracks usage and calls the merge request webhooks.

When the :use_specialized_service parameter gets used we'll be able to
call this service to avoid updating the whole merge request just to get
the spent time added.

Changelog: performance