Add total http read timeout
This introduces a global timeout for HTTP requests.
Net::ReadTimeout only measures time until the first
read but it doesn't time out if there is an ongoing
slow read. This new timeout keeps checking the time
while reading. It is however more of a workaround
to avoid replacing HTTParty which does not offer such
a global timeout.

Changelog: security