Fix LFS push mirroring for mirrors set up a long time ago
We discovered this on our own gitlab-foss -> github mirror. Where the
credentials were generated a long time ago, they don't contain an
`auth_method`. That means the method is `password`. It would be nicer
to keep the details of that within the `RemoteMirror` class, but that
would ban direct access to the `credentials` hash.

Changelog: fixed