Tweak Danger for the new changelog setup
We now only inform the user about the new setup if a changelog is
required or optional. In addition, the link to the rollout issue is now
wrapped in an inline code block. Using a regular link results in a lot
of system notes and related merge requests being created, making it
difficult to navigate the issue linked to.