Inline batched migrations only in dev env
We keep the dev env inlining for convenience for now. Otherwise,
developers would have to run the rake task whenever a data migration had
been scheduled.

For test env, we don't want this. Primarily due to executing this in the
database testing pipeline - where we don't want to inline the full
migration (it can take a long time with data from GitLab.com).

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/328755