Split feature flag removal to new template
Feature flag enabling and removal generally happen in different
milestones, but the issue template combined both actions, which made
these issues difficult to schedule accurately. This moves the feature
flag cleanup instructions to a new template, and replaces them with
an instruction to use this new template. The intent is to improve
the visibility and action on feature flag cleanup.