Use Gitaly binaries from Gitaly's build directory
When building Gitaly, it is building binaries inside of its `_build/bin`
directory, but due to legacy reasons it will also copy these binaries
into its root directory. Given that this second step is going to go away
soon, we should start to use the `_build/bin` directory.

Provide a new helper function to locate the binaries directory and
convert paths which refer to the executable to use it.