Show Deploy Boards for Environments in Folders
People may want to organize their environments by heavily relying on
folders, but, while we fetched the deploy board data, we never properly
parsed it or displayed it for users.

Here, we parse the deploy board data for children in setFolderContent,
as that is where the children are assigned to a folder when loaded.

We also need to update the check in setDeployBoard, as size does not
seem to exist on children environments. Instead, isFolder is used (which
captures the same information we need as a simple boolean).

Changelog: added