Add the ability to close a flash message in code
Until now Flash instances could only be closed by interacting
with the UI. It's often desirable to be able to dismiss a message
without user interaction and without querying the DOM.

Basically: if some code opens a Flash, it should also be able to 
close said Flash.

This update adds a simple `close` property to the container that
allows anyone with a reference to the container to close it
programmatically.

Basic use goes something like this:

```
var myFlash = createFlash(...);
// ...later, message is irrelevant now!
myFlash.close();
```