Create multiple schedules
A small MVC where we allow for multiple schedules
but gate the ~feature behind a ~"feature flag" . This merge
looks to expand the schedule code base in the most boring
way possible i.e. change the data prop from a single Object
to an Array and then allow the children components to consume
1-N schedules at a time depending on the ~"feature flag" or
how many the user has created.

This merge will also be followed by a second merge to abstract
the modals currently living inside the `oncall_schedule.vue`
component to lift them into the `oncall_schedule_wrapper.vue`
as to DRY the code up and improve performance.