Replace node-sass with sass
The node-sass project is deprecated in favor of sass, written in dart.
Apparently node-sass doesn't compile for node@16, erroring out for
community contributors and team members alike. It seems like there is
not a quick fix available for this:

https://github.com/sass/node-sass/issues/3077

Given that the project is deprecated and we do not really use it for
anything time sensitive, we can surely just replace node-sass with sass.