Add to IDE Service to Dismiss Callouts, Fetch CI
We will be using the callout mechanism to persist whether or not users
need to be told about using environments with GitLab CI. This
necessitates the ability to call graphql mutations.

We also need to be able to query the graphql API for the fully parsed CI
configuration. Fortunately, that query was already built by the pipeline
editing team, so we can re-use it here.

One thought is that it might be worth it to build our own query to only
fetch a job's environment configuration, instead of the entirety of the
pipeline, but for now re-use is better than new.