Add Store Configuration for Environment Guidance
The getAlert function searches through all available alerts (added
previously), returning the first alert where its show function returns
true, or null if none is found.

Actions to fetch the CI configuration for its given current
content (when it is being edited), as well as to dismiss the callout for
environment guidance are added here.

We parse the CI configuration during the mutation to only retain whether
or not we have:

- sucessfully parsed the CI config, and
- found an environment within the config.