Display File-Based Alerts on the WebIDE
Provides a way to display alerts attached to files within the WebIDE.

The first alert shows guidance to users who are editing their CI
configuration about environments, if we are reasonably certain they
don't know about environments.

We need to request the parsed CI configuration while the user edits it,
and accomplish this by watching the content. This watcher is
conditionally created and destroyed when the config file is active.