Improve logging in e-mail receiver processing
Previously there was little visibility in Service Desk and
reply-by-email processing. Now we add more JSON logging when mails are
successfully processed and when they fail.

To do this, we needed to refactor:

1. `Email::Receiver` and `Email::ServiceDeskReceiver` to make it
possible to extract the metadata from the mail.

2. `EmailReceiverWorker` so that`ServiceDeskEmailReceiverWorker` can
better reuse the code.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/118594

Changelog: added